math-tutor-backend/requirements.txt
fastapi
uvicorn
sympy
pydantic
python-dotenv
*.log
*.csv
openai